Patch 04/10 In hashcode related routines long changed to intptr_t, unsigned long to uintptr_t.

+++ src/integer/cl_I.h
@@ -712,7 +712,7 @@ inline sintD FN_MSD (cl_uint word)
 
 
 // Hash code.
-  extern unsigned long hashcode (const cl_I& x);
+  extern uintptr_t hashcode (const cl_I& x);
 
 
 // A fixnum (cl_FN) is an immediate integer.
